+++ builder/src/pack/mod.rs
@@ -1,2 +1,1 @@
 pub mod bundlev1;
-mod util;
