--- builder/src/pack/util.rs
@@ -1,12 +1,0 @@
-use std::io::{Seek, Write};
-use std::num::ParseIntError;
-
-pub trait WriteSeek: std::io::Write + Seek {}
-impl<T: Write + Seek> WriteSeek for T {}
-
-pub fn decode_hex(s: &str) -> Result<Vec<u8>, ParseIntError> {
-    (0..s.len())
-        .step_by(2)
-        .map(|i| u8::from_str_radix(&s[i..i + 2], 16))
-        .collect()
-}
